+++ src/components/Icon/icons/IconEmptyData/index.js
@@ -1,0 +1,17 @@
+// @flow
+import * as React from 'react';
+import { css, cx } from 'emotion';
+import { Icon, type GenericIconProps } from '../../Icon';
+import image from './icon-empty-data.svg';
+
+const styles = css`
+  width: 36px;
+  height: 45px;
+`;
+
+export const IconEmptyData = ({ className }: GenericIconProps) => (
+  <Icon
+    className={cx(styles, className)}
+    glyph={image}
+  />
+);