+++ src/components/Icon/index.js
@@ -27,6 +27,7 @@ export * from './icons/IconDelete';
 export * from './icons/IconDocumentCode';
 export * from './icons/IconDownload';
 export * from './icons/IconEdit';
+export * from './icons/IconEmptyData';
 export * from './icons/IconEyeClosed';
 export * from './icons/IconEyeOpened';
 export * from './icons/IconFile';