+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add new icon `IconEmptyData`;
+
+- Update component `NonIdealState`. Add new prop `imageClassName`;
+
+- Update component `Table`:
+    
+    - Remove props: `codeClassName`, `onCodeRowClick`, `codeRowKey`;
+    
+    - Update no data state. Use new icon `IconEmptyData`.
+
 ## [0.40.1] - 2021-08-02
 
 - Fix svg, optimize svg